fixed "cannot export stores in page.tsx"
