+++ src/app/learnmoves/data/learn/MovementGroup.tsx
@@ -24,9 +24,20 @@ import { Form, SubmitHandler, useForm } from 'react-hook-form'
 
 import { RenderEditButton, RenderRedDeleteButton } from '../../_components/Svgs'
 import { RenderAddButton } from '../../_components/Svgs'
-import { useDataLearnStore } from './page'
 import { MovementKeys, MovementType } from './pagetypes'
 import { RenderHearts } from './RenderHearts'
+import { create } from 'zustand'
+// ----------------------store-----------------------------
+export interface DataLearnState {
+  //note key for isEditing is actually a number from an index array fnc. however in js all keys are strings.
+  isEditing: { [key: string]: boolean } | null
+  setIsEditing: (val: { [key: string]: boolean }) => void
+}
+
+export const useDataLearnStore = create<DataLearnState>()((set) => ({
+  isEditing: null,
+  setIsEditing: (val) => set(() => ({ isEditing: val })),
+}))
 
 //-------------local types------------------
 //input types for react-hook-form
@@ -122,17 +133,17 @@ export default function RenderMovementGroup({
     //-----makes a defaults if none found to handle edge cases----
     //do not have a default for the last movementgroup as it's just a transition loop to repeat and doesnt have positions
     position = indexNumber !== localMovements.length - 1 &&
-      makeDefaultPosition({
-        displayName: 'new-position',
-      }),
+    makeDefaultPosition({
+      displayName: 'new-position',
+    }),
     //doesn't make a transitionobj for the first pos, as nothing to transition from
     transition = indexNumber !== 0 &&
-      makeDefaultTransition({
-        displayName: 'new-transition',
-        from: localMovements[indexNumber - 1].positionId || makePositionId(),
-        to: position ? position.positionId : makePositionId(),
-        transitionId: makeTransitionId(),
-      }),
+    makeDefaultTransition({
+      displayName: 'new-transition',
+      from: localMovements[indexNumber - 1].positionId || makePositionId(),
+      to: position ? position.positionId : makePositionId(),
+      transitionId: makeTransitionId(),
+    }),
   } = getPositionAndTransition(movement, move)
   //----------------use effect-----------
   //makes sure has access to local storage
